Type change to avoid Windows compilation issues

+++ src/lib/ndpi_utils.c
@@ -1748,7 +1748,7 @@ const char* ndpi_http_method2str(ndpi_http_method m) {
 
 /* ******************************************************************** */
 
-ndpi_http_method ndpi_http_str2method(const char* method, ssize_t method_len) {
+ndpi_http_method ndpi_http_str2method(const char* method, u_int16_t method_len) {
   if(!method || method_len < 3)
     return(NDPI_HTTP_METHOD_UNKNOWN);
